Crate authzen_opa_core

Structs

AccountSessionFields

DeleteDocument

DeleteDocumentParams

GetConfig

GetConfigParams

GetDocument

GetDocumentParams

GetDocumentResult

GetStatus

GetStatusParams

Health

HealthParams

OPAClient

OPAPolicyASTNodeRef

OPAPolicyASTNodeRefRef

OPAQuery

OPAQueryConfig

OPAQueryInput

OPAQueryResult

OPAResult

OPATxEntity

OPATxEntityFullmongodb-tx-cache-backend

OPAType

PatchDocument

uses json patch format

UpsertDocument

UpsertDocumentParams

Enums

CowSlice

DisableAuthorization

use this type in places where authorization disabling needs to be very clear in all contexts e.g. as a function parameter, where both the function and the function caller should be unambiguously aware of what the parameter they are using / passing means

OPAPolicyASTNode

OPAPolicyPathNode

OPAQueryInputAction

Constants

DEFAULT_TIMEOUT_SECONDS

DEFAULT_TTL_SECONDSmongodb-tx-cache-backend

TTL_INDEX_NAMEmongodb-tx-cache-backend

Traits

AllOPAIdentifiable

AuthzService

AuthzServiceEntity

OPAContext

OPAIdentifiable

OPAIdentifiableEntity

OPAKey

OPATxCacheClient

OPATxCacheContext

Functions

initialize_ttl_indexmongodb-tx-cache-backend

query_all

Type Definitions

OPADataCacheUpsert

OPAMongoCollectionmongodb-tx-cache-backend

OPAMongoCollectionValuemongodb-tx-cache-backend